dom/media/mediasource/test/test_BufferingWait.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: Don't get stuck buffering for too long when we have frames to show</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  ok(true, "Receive a sourceopen event");
  ms.addEventListener("sourceopen", () => ok(false, "No more sourceopen"));
  const sb = ms.addSourceBuffer("video/webm");
  ok(sb, "Create a SourceBuffer");

  const arrayBuffer = await fetchWithXHR("seek.webm");
  sb.addEventListener("error", e => {
    ok(false, "Got Error: " + e);
    SimpleTest.finish();
  });
  await loadSegment(sb, new Uint8Array(arrayBuffer, 0, 318));
  await loadSegment(sb, new Uint8Array(arrayBuffer, 318, 25523 - 318));
  await loadSegment(sb, new Uint8Array(arrayBuffer, 25523, 46712 - 25523));
  /* Note - Missing |46712, 67833 - 46712| segment here corresponding to (0.8, 1.2] */
  /* Note - Missing |67833, 88966 - 67833| segment here corresponding to (1.2, 1.6]  */
  await loadSegment(sb, new Uint8Array(arrayBuffer, 88966));
  // 0.767 is the time of the last video sample +- 40ms.
  info("Playing video. It should play for a bit, then fire 'waiting'");
  v.play();
  await waitUntilTime(v, .767 - 0.04);
  const firstStop = Date.now();
  await loadSegment(sb, new Uint8Array(arrayBuffer, 46712, 67833 - 46712));
  await waitUntilTime(v, 1.167 - 0.04);
  const waitDuration = (Date.now() - firstStop) / 1000;
  ok(waitDuration < 15, `Should not spend inordinate amount of time buffering: ${waitDuration}`);
  SimpleTest.finish();
  /* If we allow the rest of the stream to be played, we get stuck at
     around 2s. See bug 1093133.
  await once(v, "ended");
  SimpleTest.finish();
  await loadSegment(sb, new Uint8Array(arrayBuffer, 67833, 88966 - 67833));
  */
});
</script>
</pre>
</body>
</html>
